suzerain: add test for critical FFTW patch
Pulled out the marked failures committed not long ago; it is possible to have
these test cases pass, but it requires a patch I submitted to the FFTW folks on
Friday/today.  Waiting to hear back from them about its acceptance.  If it goes
in, then I leave the checks in place.  If it does not, then I account for it
within my own code.


git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@7701 0bd85f0b-c151-43b9-a1e9-27231104551f