suzerain: #1573 check discrete symmetry (breaks tests)
We store the general banded mass matrix for the Galerkin L2
operators though it is technically symmetric banded.  This is
because higher derivative matrices are not symmetric and we
want them to have compatible storage.

Include a check to ensure our mass matrix is discretely symmetric,
even in the presence of floating point error.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@19381 0bd85f0b-c151-43b9-a1e9-27231104551f