suzerain: #1908 custom reduction function draft
I want to hide the communications overhead of looking for
either signal receipt or wall time completion.  The first step
in doing so is that provide a hook where the Allreduce used to
compute a stable time step can be rigged to perform these operations
too.

Augment INonlinearOperator to have it return one or more time stepping
criteria.

Add a DeltaTReducer (with defaults) that computes the minimum over all
returned values (with NaNs being propagated correctly).

Change all NonlinearOperator instances to match the new signature.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@23145 0bd85f0b-c151-43b9-a1e9-27231104551f