suzerain: #2063 remove N.O.Isothermal + Dox
The BsplineMassOperatorIsothermal is getting updated with
the new sampling logic for bar_f, bar_f_dot_u, and bar_qb.
I don't want to write this in NonlinearOperatorIsothermal
and support sampling from two locations.  That and
NonlinearOperatorIsothermal has been deadweight for awhile.
Remove it.

Update Doxygen to reflect the communication amongs the remaining
operator classes.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@25606 0bd85f0b-c151-43b9-a1e9-27231104551f