suzerain: #2215 decouple bulk_rho from solve
See comments in revision.  The coupling between the energy and density
equation's integral constraints was interfering with enforcing the
isothermal condition.  This revision (more or less) preserves the
behavior of the explicit channel while making things about an order of
magnitude better (measured in the wall constraint on rho and rhoe) for
the implicit case.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@28041 0bd85f0b-c151-43b9-a1e9-27231104551f