suzerain: #2515 correct grad div m
Similar to r23801, I had an incorrect version of grad div vector
within the discrete operators.  This caused noticeable linearization
errors in rhou (based on nu) and in rhow (based on nu) equations.

Fixed in writeup and code. Tests isolating the impact of the
linearization error due to these terms now pass.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@28304 0bd85f0b-c151-43b9-a1e9-27231104551f