suzerain: #2321 fix linear op on high modes
The changes from r28403 avoided applying the implicit linear
operator on high wavenumbers.  I made a copy'n'paste mistake
from diffwave.c and eliminated all wavenumbers not used in
the derivative of the fields instead of all wavenumbers
not used in the fields themselves.

That's a fancy way of saying I made a stupid off-by-one error.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@28415 0bd85f0b-c151-43b9-a1e9-27231104551f