suzerain: #2383 leave Nyquist alone on store
Previously, the act of storing coefficients to disk would zero out the
imaginary part of Nyquist modes.  This commit stops that-- data is
saved out to disk as-is in memory.

Previously, nonlinear operator application would leave non-zero
imaginary parts for the Nyquist modes.  This commit stops that--
nonlinear operator application no longer bleeds Nyquist information
into the remainder of the code.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@31309 0bd85f0b-c151-43b9-a1e9-27231104551f