suzerain: #2579 move rho\([Euvw]\) to rho_\1
THIS IS A BACKWARDS-COMPATIBILITY BREAKING CHANGE!

That is, make the following changes
  rhoe -> rho_E
  rhou -> rho_u
  rhov -> rho_v
  rhow -> rho_w
everywhere except in the manufactured solution API (as
that has been published and the checksum is useful).

Updates have been made to the shipped fields using postproc/rename_hdf5.py:
  ./postproc/rename_hdf5.py rhoe rho_E */*.h5
  ./postproc/rename_hdf5.py rhou rho_u */*.h5
  ./postproc/rename_hdf5.py rhov rho_v */*.h5
  ./postproc/rename_hdf5.py rhow rho_w */*.h5
Similar statements should permit updating any existing legacy data.

Reason for breaking compatibility is the complexity of the change
would be too high to permit legacy data to co-exist (especially in
samples) which the utility to upgrade is simple.  Necessary to get
this rename done before the inconsistency goes off into the wild.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@34120 0bd85f0b-c151-43b9-a1e9-27231104551f