suzerain: #2480 towards general number of fields
Much of this is plugging in suzerain::ndx instead of support::field

Not 100% happy, but baby steps towards getting a variable
number of fields into the general support routines.
Intent is for the driver to house and propagate the field
information down the line.

This commit is dubious in that
  A) It changed a lot of things
  B) I'm not testing it before I commit because I'm on the wrong system

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@34290 0bd85f0b-c151-43b9-a1e9-27231104551f