suzerain: consolidate nonlinear right hand side
With common_block/navier_stokes now isolated from the previous commit,
all nonlinear computations are moved to nonlinear_operator.  That one
class is able to deal with the (currently) two separate linearizations.

The nonlinear operator from hybrid_operator is removed.

The nonlinear operator from explicit_operator is removed.  The
file is renamed to isothermal_mass_operator as that's all it
still contains.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@38206 0bd85f0b-c151-43b9-a1e9-27231104551f