suzerain: add RPATH to installed libsuzerain
Nick was installing in a prefix not on the LD_LIBRARY_PATH,
and the binary was unable to find the associated libsuzerain.
This commit adds an explicit RPATH flag to the build process
for libsuzerain so that installed applications from apps/
will always be able to find the corresponding library.


git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@38768 0bd85f0b-c151-43b9-a1e9-27231104551f