suzerain: #2399 refactor integral constraints
Implementation theoretically permits constraining any combination of
rho, rho_u, rho_v, rho_w, and rho_E.  Testing shows all of the
following working:
- DISABLED
- rho
- rho_u
- rho_E
- rho, rho_u
- rho, rho_E
but anything combining both rho_u and rho_E is hosed.  That is,
- rho_u, rho_E
- rho, rho_u, rho_E
both bomb.

I have confirmed that bug is a regression from the last implementation
and DOES NOT indicate a bug in the math or physics.  Or, at least for
some values (e.g. --bulk_rho=0.75 --bulk_rho_u=1 --bulk_rho_E=1.95)
it does not indicate a bug in the math.  The physics may be out to lunch.

Regardless, I'm committing this as it's been too long in the making.
Getting nervious with this outside the repo.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@39501 0bd85f0b-c151-43b9-a1e9-27231104551f