suzerain: #2477 remove lowstorage namespace
Work on #2477 suggests that having nonlinear_operator's
apply_operator method take a lowstorage::method_interface
is a good idea, but nonlinear_operator is in a higher
namespace.  Make suzerain::timestepper always be low-storage
and therefore remove the sub-namespace lowstorage once-and-for-all.

Only naming casualty is lowstorage_timecontroller is now timestepper::controller.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@40092 0bd85f0b-c151-43b9-a1e9-27231104551f