suzerain: #2983 document progress, remove stubs
Rips out test code that I'd used over the last couple of days
to attempt to get viscous subsonic outflow conditions working
on the nonreflecting upper boundary.

Tosses out a warning message on startup if the problem is
known to occur.  Hopefully this lets someone find the issue
down the line.

Updates the manual.nrbc/ticket2399.sh test case so people know
it is expected to fail until #2983 and #2957 are resolved.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@42328 0bd85f0b-c151-43b9-a1e9-27231104551f