suzerain: #2956 first_dt O(eps) bugfix
Hit a pathological edge case in practice where first_dt was nearly
an integer multiple of every_dt, and the difference from the
integer was O(machine epsilon) relative to the current simulation
time.  This caused an invariant about e->next_t to be broken,
caused possible_dt == 0, and brought down the simulation.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@43058 0bd85f0b-c151-43b9-a1e9-27231104551f