suzerain: remove ifdef HAVE_CONFIG_H
I wanted to flip
  #ifdef HAVE_CONFIG_H
  #include <suzerain/config.h>
  #endif
to
  #ifdef HAVE_CONFIG_H
  #include "config.h"
  #endif
within suzerain/**/*.c suzerain/**/*.cpp but my use of
Automake's nostdinc option precluded it.  I tried whacking
nostdinc along with that flip, and faced mutiny on part of
the codebase.

Then I remembered that the suzerain/suzerain-config.h
header is intended to handle these details, and that it
gets pulled in everytime that common.{h,hpp} is used.

So, I'm removing the needless #ifdef HAVE_CONFIG_H
boilerplate throughout the library.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@44241 0bd85f0b-c151-43b9-a1e9-27231104551f