suzerain: #3058 no bulk_rho_{,_u,_E} means NaN
Tweaks the load time behavior so that removing the bulk constraint
targets effectively disables them on start up.  Not pretty, but
effective and quite backwards compatible.  Notice that the new
NaN values will be written back to disk so that the slight of hand
is detectable.  And that we WARN0 when the trick is played.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@44242 0bd85f0b-c151-43b9-a1e9-27231104551f