suzerain: #2983 #3070 inviscid freestream boundary
Adds a boolean that tracks whether or not the computation should
be "locallyviscous".  Always true unless we're at the freestream
boundary on a one-sided grid.  When !locallyviscous, local mu
and lambda are set to be zero.  Notice, grad_mu and grad_lambda
are not modified (hence, "locally").

This tweak is meant to produce an inviscid-like upper boundary
that neglects second order derivatives as suggested by Poinsot
and Lele.  For simplicity, distinguishing between inflow and
outflow is nto done.

Performed within reference gathering and RHS computations.
In particular, doing this within reference gathering should, I think,
disable the viscous action within the linear implicit operator as well.
Tests on lin.abserr show FIXME

Tracking that the freestream is inviscid is /NOT/ done within the
sampling routines as it perturbs the statistics badly near the
freestream in a way they do not see during the computation.  That is,
one point off the freestream grad_mu, etc. is advanced in time as
if the freestream is viscous.  It would be nuts to change the
Reynolds-averaged viscosity going into the restart file at the freestream
as that information is used to compute wall-normal derivatives one off
the boundary.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@44772 0bd85f0b-c151-43b9-a1e9-27231104551f