suzerain: default SLURM outputs OK
The defaults per the manual are just fine:

By default both standard output and standard error are directed to the same file. For job arrays, the default file name is "slurm-%A_%a.out", "%A" is replaced by the job ID and "%a" with the array index. For other jobs, the default file name is "slurm-%j.out", where the "%j" is replaced by the job ID.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@44926 0bd85f0b-c151-43b9-a1e9-27231104551f