suzerain: #3016 dont require delta1 for delta
Prior versions of this logic required a sane delta1 for
delta to be found.  As that apparently won't happen
for many of the runs, fall back to not requiring a good
delta1 to compute delta.  More likely to produce false
results near the wall, but at least will give back some
answer.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@45036 0bd85f0b-c151-43b9-a1e9-27231104551f