suzerain: #3109 move 1/chi out of treatment members
Want to use this as a superclass for operator_hybrid_isothermal,
and the first step is making it so that subclasses can use
non-scaled versions of these operators.  The scaling factor
1/chi application cost is now hidden in the wavenumber computations.

Notice compute_giles_matrices_upper is now virtual, so that
we can later hook it by operator_hybrid_isothermal.

git-svn-id: https://svn.ices.utexas.edu/repos/pecos/suzerain/trunk@45200 0bd85f0b-c151-43b9-a1e9-27231104551f